+++ client/package.json
@@ -5,8 +5,11 @@
   "private": true,
   "proxy": "http://localhost:3001/",
   "dependencies": {
+    "axios": "^0.19.0",
+    "cheerio": "^1.0.0-rc.3",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
+    "react-router-dom": "^5.0.1",
     "react-scripts": "^2.1.1"
   },
   "scripts": {
