+++ client/public/index.html
@@ -17,10 +17,10 @@
 
     <script src="./embed.js"></script>
 
-    <script>
+    <!-- <script>
         var options = {};
         Botkit.boot(options);
-    </script>
+    </script> -->
 
 
 <!-- App Page -->
