+++ client/src/App.js
@@ -12,21 +12,21 @@ class App extends Component {
     <div>
       <div id='splash'>
           <div id='bars'>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
-            <div class='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
+            <div className='bar'></div>
             <img id="Apptitle" src="https://i.ibb.co/JjXYSBh/logo.png"></img>
           </div>
