+++ controllers/botkitcontroller.js
@@ -1,7 +1,7 @@
-let { Botkit } = require('botkit');
+// let { Botkit } = require('botkit');
 
-const controller = new Botkit();
+// const controller = new Botkit();
 
-controller.hears('hello','direct_message', function(bot, message) {
-    bot.reply(message,'Hello yourself!');
-});
+// controller.hears('hello','direct_message', function(bot, message) {
+//     bot.reply(message,'Hello yourself!');
+// });
