+++ package.json
@@ -20,6 +20,7 @@
   },
   "dependencies": {
     "axios": "^0.19.0",
+    "botkit": "^4.0.2",
     "cheerio": "^1.0.0-rc.3",
     "express": "^4.16.3",
     "express-favicon": "^2.0.1",
